server/requirements.txt
flask
flask-cors
torch
torchvision
numpy
pyngrok
python-dotenv